AmbientIndication: Fix music ticker being black on AOD when using light wallpaper
In the current implementation, if we set a light wallpaper, the
text color of the music ticker is set to black, which is correct
for the keyguard -- but not for AOD/Ambient display.

This commit uses a similar check as is used in the AOSP statusbar
to distinguish whether a light or dark wallpaper is in use.
Of course, on AOD/ambient display the text color must always be white.

Signed-off-by: spezi77 <spezi7713@gmx.net>