SystemUI: Prevent whitelist packages from appearing in privacy dialog
Previously, we only block whitelist packages from triggering privacy
indicator, which means when other packages trigger the indicator, the
whitelist packages can still be seen in the privacy dialog. This can
be confusing as the privacy type showed in indicator can be mismatch
with that showed in dialog.

Fix it by applying whitelist to privacy dialog too.

Change-Id: I4dadb73eaa525e5b3ef1c663654ef00a353866cd
Signed-off-by: LibXZR <i@xzr.moe>